Refactor the auto class, add positional arguments
