+++ mgtbench/__init__.py
@@ -1,8 +1,9 @@
+from . import auto
 from .loading import (
     dataloader, dataloader_attribution
     )
 from .methods import (
-    auto, detectgpt,gptzero,IntrinsicDim,metric_based,supervised
+    detectgpt,gptzero,IntrinsicDim,metric_based,supervised
     )
 
-from .methods.auto import AutoDetector, MetricBasedDetector, ModelBasedDetector
+from .auto import AutoDetector, MetricBasedDetector, ModelBasedDetector, AutoExperiment
