+++ mgtbench/methods/__init__.py
@@ -2,3 +2,6 @@ from .metric_based import  get_ll, get_rank, get_entropy, get_rank_GLTR, run_thr
 from .gptzero import run_gptzero_experiment, GPTZeroAPI
 from .detectgpt import run_perturbation_experiments
 from .supervised import run_supervised_experiment
+
+
+from .metric_based import LLDetector, RankDetector
